requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pillow>=9.5.0
requests>=2.28.0
